Rename to DeviceListenerCurrentDevice
